refactor: separate mem test from lib.rs
