+++ crates/core/src/lib.rs
@@ -1,47 +1,15 @@
 #![no_std]
 
-use core::fmt::Arguments; 
+mod test;
 
-pub fn run_test<M: Mem, L: Logger>(logger: &L, mem: &M, region: MemoryRegion) {
+use core::fmt::Arguments;
+use crate::test::marchc;
+
+pub fn run_test<M: Mem, L: Logger>(logger: &L, mem: &M, region: &MemoryRegion) {
     logger.log(format_args!("Checking region {:?}", region));
-    let offset_region = mem.parse(region);
 
-    for addr in offset_region.start..offset_region.end {
-        if !mem.check(addr) {
-            continue;
-        }
-        mem.write(addr, 0);
-    }
 
-    for addr in offset_region.start..offset_region.end {
-        if !mem.check(addr) {
-            continue;
-        }
-        if mem.read(addr) != 0 {
-            panic!("Test failed at address 1 {:?}", addr);
-        }
-        mem.write(addr, 1);
-    }
-
-    for addr in (offset_region.start..offset_region.end).rev() {
-        if !mem.check(addr) {
-            continue;
-        }
-        if mem.read(addr) != 1 {
-            panic!("Test failed at address 2 {:?}", addr);
-        }
-        mem.write(addr, 0);
-    }
-
-    for addr in offset_region.start..offset_region.end {
-        if !mem.check(addr) {
-            continue;
-        }
-        if mem.read(addr) != 0 {
-            panic!("Test failed at address 3 {:?}", addr);
-        }
-        mem.write(addr, 1);
-    } 
+    marchc::run_march_c(mem, region);
 }
 
 #[derive(Debug)]
@@ -51,7 +19,7 @@ pub struct MemoryRegion {
 }
 
 pub trait Mem {
-    fn parse(&self, region: MemoryRegion) -> MemoryRegion;
+    fn parse(&self, region: &MemoryRegion) -> MemoryRegion;
     fn check(&self, addr: u64) -> bool;
     fn read(&self, addr: u64) -> u64;
     fn write(&self, addr: u64, value: u64);
@@ -60,4 +28,3 @@ pub trait Mem {
 pub trait Logger {
     fn log(&self, message: Arguments<'_>);
 }
-
